# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
  isActive?: boolean;
  createdAt: Date;
}