# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
  /** @minLength 6 */
  password?: string;
}